+++ application/src/main/java/common/bolts/transactional/sl/SLBolt_sstore.java
@@ -79,7 +79,7 @@ public class SLBolt_sstore extends SLBolt_LA {
         }
     }
 
-    public void start_evaluate(int thread_Id, long mark_ID, int num_events) throws InterruptedException, BrokenBarrierException {
+    public void start_evaluate(int thread_Id, double mark_ID, int num_events) throws InterruptedException, BrokenBarrierException {
         SOURCE_CONTROL.getInstance().preStateAccessBarrier(thread_Id);//sync for all threads to come to this line to ensure chains are constructed for the current batch.
         LA_RESETALL(transactionManager, thread_Id);
         // add bid_array for events
@@ -119,8 +119,8 @@ public class SLBolt_sstore extends SLBolt_LA {
     }
 
     @Override
-    protected void PRE_TXN_PROCESS(long _bid, long timestamp) throws DatabaseException, InterruptedException {
-        for (long i = _bid; i < _bid + combo_bid_size; i++) {
+    protected void PRE_TXN_PROCESS(double _bid, long timestamp) throws DatabaseException, InterruptedException {
+        for (double i = _bid; i < _bid + combo_bid_size; i++) {
 //            System.out.println("thread: "+thread_Id+", event_id: "+_bid);
             TxnEvent event = (TxnEvent) input_event;
             GlobalSorter.addEvent(event);
@@ -128,7 +128,7 @@ public class SLBolt_sstore extends SLBolt_LA {
     }
 
     @Override
-    protected void LAL_PROCESS(long _bid) throws DatabaseException {
+    protected void LAL_PROCESS(double _bid) throws DatabaseException {
         txn_context[0] = new TxnContext(thread_Id, this.fid, _bid);
         TxnEvent event = (TxnEvent) input_event;
         int _pid = (event).getPid();
