+++ state-engine/src/main/java/content/ToContent.java
@@ -12,12 +12,12 @@ public abstract class ToContent implements Content {
     }
 
     @Override
-    public void SetTimestamp(long timestamp) {
+    public void SetTimestamp(double timestamp) {
         throw new UnsupportedOperationException();
     }
 
     @Override
-    public long GetTimestamp() {
+    public double GetTimestamp() {
         throw new UnsupportedOperationException();
     }
 
@@ -57,7 +57,7 @@ public abstract class ToContent implements Content {
     }
 
     @Override
-    public long GetLWM() {
+    public double GetLWM() {
         throw new UnsupportedOperationException();
     }
 
@@ -96,7 +96,7 @@ public abstract class ToContent implements Content {
     }
 
     @Override
-    public void AddLWM(long ts) {
+    public void AddLWM(double ts) {
         throw new UnsupportedOperationException();
     }
