+++ state-engine/src/main/java/scheduler/impl/op/structured/OPSScheduler.java
@@ -35,7 +35,7 @@ public class OPSScheduler<Context extends OPSContext> extends OPScheduler<Contex
     }
 
     @Override
-    public void start_evaluation(Context context, long mark_ID, int num_events) {
+    public void start_evaluation(Context context, double mark_ID, int num_events) {
         int threadId = context.thisThreadId;
 
         INITIALIZE(context);
@@ -99,7 +99,7 @@ public class OPSScheduler<Context extends OPSContext> extends OPScheduler<Contex
     }
 
     @Override
-    public void PROCESS(Context context, long mark_ID) {
+    public void PROCESS(Context context, double mark_ID) {
         int cnt = 0;
         int batch_size = 100;//TODO;
         int threadId = context.thisThreadId;
